Fix NanBug and Payment Could Run
Fix NanBug and Payment Could Run

+++ tessar-netty-server/src/main/java/vova/domain/payment/PayMentWeek.java
@@ -10,7 +10,8 @@ public class PayMentWeek extends PayAllShow {
     public PayMentWeek() {
     }
 
-    public PayMentWeek(int id, Date dateID, String cId, String gId, String sId, long newAddNum, long newAddPayPlayerNum, long newAddPayAllMoney, float newAddPlayerARPPU, float newAddPayPlayerARPPU, long newPayPlayerNum, long newPayAllMoney, float averageNewPayMoney, long activeNum, long payPlayerNum, long todayAllPayMoney, float activeARPPU, float payPlayerARPPU, long allMoney) {
+    public PayMentWeek(int id, Date dateID, String cId, String gId, String sId, long newAddNum, long newAddPayPlayerNum, float newAddPayAllMoney, float newAddPlayerARPPU, float newAddPayPlayerARPPU, long newPayPlayerNum, float newPayAllMoney, float averageNewPayMoney, long activeNum, long payPlayerNum, float todayAllPayMoney, float activeARPPU, float payPlayerARPPU, float allMoney) {
         super(id, dateID, cId, gId, sId, newAddNum, newAddPayPlayerNum, newAddPayAllMoney, newAddPlayerARPPU, newAddPayPlayerARPPU, newPayPlayerNum, newPayAllMoney, averageNewPayMoney, activeNum, payPlayerNum, todayAllPayMoney, activeARPPU, payPlayerARPPU, allMoney);
     }
+
 }
